#52: Query tool and Bulldozer now function. Fixed #62.

+++ Assets/Scripts/Interfaces/IApproval.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 
 public interface IApproval
 {
-    float GetApprovalScore(IReadOnlyDictionary<Vector3Int, Building> cityTiles);
+    double GetApprovalScore(IReadOnlyDictionary<Vector3Int, Building> cityTiles);
 }
